feat: increase node index from u32 to u64
u32 is already almost fully utilzied on mainnet.

This is just a temporary fix until we replace the index with
proper newtypes.

+++ crates/merkle-tree/src/transaction.rs
@@ -28,11 +28,11 @@ impl Default for TransactionOrEventTree {
 struct NullStorage;
 
 impl crate::storage::Storage for NullStorage {
-    fn get(&self, _: u32) -> anyhow::Result<Option<StoredNode>> {
+    fn get(&self, _: u64) -> anyhow::Result<Option<StoredNode>> {
         Ok(None)
     }
 
-    fn hash(&self, _: u32) -> anyhow::Result<Option<Felt>> {
+    fn hash(&self, _: u64) -> anyhow::Result<Option<Felt>> {
         Ok(None)
     }
